Split into core (lib) and console app
