+++ ArchiveUnpacker.CommandLine/ArchiveUnpacker.CommandLine.csproj
@@ -1,0 +1,31 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net472</TargetFramework>
+        <LangVersion>7.3</LangVersion>
+        <AssemblyName>Macaron</AssemblyName>
+        <Title>Macaron</Title>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="CommandLineParser" Version="2.4.3" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\ArchiveUnpacker.Core\ArchiveUnpacker.Core.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <Compile Remove="Extracted\**" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <EmbeddedResource Remove="Extracted\**" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <None Remove="Extracted\**" />
+    </ItemGroup>
+
+</Project>
