+++ ArchiveUnpacker.CommandLine/CommandLineParsing/ExtractOptions.cs
@@ -1,0 +1,11 @@
+using CommandLine;
+
+namespace ArchiveUnpacker.CommandLine.CommandLineParsing
+{
+    [Verb("extract", HelpText = "Extract all files from an archive or game folder")]
+    public class ExtractOptions
+    {
+        [Value(0, MetaName = "Game directory", HelpText = "The root directory of the game", Required = true)]
+        public string Directory { get; set; }
+    }
+}
