+++ ArchiveUnpacker.Core/Framework/Exceptions/InvalidMagicException.cs
@@ -1,0 +1,7 @@
+using System;
+
+namespace ArchiveUnpacker.Core.Framework.Exceptions
+{
+    // Should prob make this better
+    public class InvalidMagicException : Exception { }
+}
