+++ ArchiveUnpacker.Core/Framework/IExtractableFile.cs
@@ -1,0 +1,10 @@
+using System.IO;
+
+namespace ArchiveUnpacker.Core.Framework
+{
+    public interface IExtractableFile
+    {
+        string Path { get; }
+        void WriteToStream(Stream writeTo);
+    }
+}
