+++ ArchiveUnpacker.Core/Framework/IUnpacker.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+
+namespace ArchiveUnpacker.Core.Framework
+{
+    public interface IUnpacker
+    {
+        IEnumerable<IExtractableFile> LoadFiles(string gameDirectory);
+        IEnumerable<IExtractableFile> LoadFilesFromArchive(string inputArchive);
+    }
+}
