+++ ArchiveUnpacker.Core/Utils/Pickle/PickleReader.cs
@@ -1,0 +1,13 @@
+using System.IO;
+
+namespace ArchiveUnpacker.Core.Utils.Pickle
+{
+    public static class PickleReader
+    {
+        public static object ReadFromStream(Stream s)
+        {
+            var state = new PickleState();
+            return state.ReadFromStream(s);
+        }
+    }
+}
