--- ArchiveUnpacker/ArchiveUnpacker.csproj
@@ -1,28 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>net472</TargetFramework>
-        <LangVersion>7.3</LangVersion>
-        <AssemblyName>Macaron</AssemblyName>
-        <Title>Macaron</Title>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <PackageReference Include="CommandLineParser" Version="2.4.3" />
-      <PackageReference Include="TriggersTools.Windows.Resources" Version="1.0.2-build-000011" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <Compile Remove="Extracted\**" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <EmbeddedResource Remove="Extracted\**" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <None Remove="Extracted\**" />
-    </ItemGroup>
-
-</Project>
