--- ArchiveUnpacker/Framework/Exceptions/InvalidMagicException.cs
@@ -1,7 +1,0 @@
-using System;
-
-namespace ArchiveUnpacker.Framework.Exceptions
-{
-    // Should prob make this better
-    public class InvalidMagicException : Exception { }
-}
