--- ArchiveUnpacker/Framework/IExtractableFile.cs
@@ -1,10 +1,0 @@
-using System.IO;
-
-namespace ArchiveUnpacker.Framework
-{
-    public interface IExtractableFile
-    {
-        string Path { get; }
-        void WriteToStream(Stream writeTo);
-    }
-}
